reaction-video-venv-requirements.txt
accelerate
torch
torchaudio
torchsummary
torchvision
torchmetrics[image]
transformers
google-api-python-client
yt-dlp
matplotlib
seaborn
pandas
numpy
scipy
scikit-learn
scikit-image
opencv-python
diffusers
# video processing
scenedetect[opencv]
# LDL metrics
python-ldl

# # antlr4-python3-runtime==4.9.3
# # anyio==4.6.2.post1
# # argon2-cffi==23.1.0
# # argon2-cffi-bindings==21.2.0
# # arrow==1.3.0
# # asttokens==2.4.1
# # async-lru==2.0.4
# # attrs==24.2.0
# # babel==2.16.0
# # beautifulsoup4==4.12.3
# # bleach==6.2.0
# # boto3==1.26.8
# # botocore==1.29.8
# # click==8.1.7
# # comm==0.2.2
# # contourpy==1.0.6
# # cycler==0.11.0
# # debugpy==1.8.7
# # decorator==4.4.2
# # defusedxml==0.7.1
# diffusers
# # exceptiongroup==1.2.2
# # executing==2.1.0
# # face-recognition==1.3.0
# # face_recognition_models==0.3.0
# # facenet-pytorch==2.6.0
# # fastjsonschema==2.20.0
# # h11==0.14.0
# # hpack==4.0.0
# # httpcore==1.0.6
# # httpx==0.27.2
# # huggingface-hub==0.25.0
# # imageio==2.36.0
# # imageio-ffmpeg==0.5.1
# # importlib-metadata==5.0.0
# # importlib_resources==6.4.5
# # ipykernel==6.29.5
# # ipython==8.18.1
# # ipywidgets==8.1.5
# # isoduration==20.11.0
# # jaraco.classes==3.2.3
# # jedi==0.19.1
# # jeepney==0.8.0
# # jmespath==1.0.1
# # json5==0.9.25
# # jsonpointer==3.0.0
# # jsonschema==4.23.0
# # jsonschema-specifications==2024.10.1
# jupyter==1.1.1
# jupyter-console==6.6.3
# jupyter-events==0.10.0
# jupyter-lsp==2.2.5
# jupyter_client==8.6.3
# jupyter_core==5.7.2
# jupyter_server==2.14.2
# jupyter_server_terminals==0.5.3
# jupyterlab==4.2.5
# jupyterlab_pygments==0.3.0
# jupyterlab_server==2.27.3
# jupyterlab_widgets==3.0.13
# # keyring==23.11.0
# # kiwisolver==1.4.4
# # lazy_loader==0.4
# # matplotlib==3.9.2
# # matplotlib-inline==0.1.7
# # mistune==3.0.2
# # mock==4.0.3
# # more-itertools==9.0.0
# # moviepy==1.0.3
# # nbclient==0.10.0
# # nbconvert==7.16.4
# # nbformat==5.10.4
# # nda-tools==0.2.20
# # nest-asyncio==1.6.0
# # networkx==2.8.8
# # notebook==7.2.2
# # notebook_shim==0.2.4
# # numpy==1.26.0
# # nvidia-cublas-cu12==12.1.3.1
# # nvidia-cuda-cupti-cu12==12.1.105
# # nvidia-cuda-nvrtc-cu12==12.1.105
# # nvidia-cuda-runtime-cu12==12.1.105
# # nvidia-cudnn-cu12==8.9.2.26
# # nvidia-cufft-cu12==11.0.2.54
# # nvidia-curand-cu12==10.3.2.106
# # nvidia-cusolver-cu12==11.4.5.107
# # nvidia-cusparse-cu12==12.1.0.106
# # nvidia-nccl-cu12==2.19.3
# # nvidia-nvjitlink-cu12==12.6.77
# # nvidia-nvtx-cu12==12.1.105
# # omegaconf==2.3.0
# # opencv-python==4.10.0.84
# # overrides==7.7.0
# # packaging==24.1
# # pandas==1.5.2
# # pandocfilters==1.5.1
# # parso==0.8.4
# # pexpect==4.9.0
# # pillow==10.2.0
# # platformdirs==4.3.6
# # proglog==0.1.10
# # prometheus_client==0.21.0
# # prompt_toolkit==3.0.48
# # psutil==6.1.0
# # ptyprocess==0.7.0
# # pure_eval==0.2.3
# # Pygments==2.18.0
# # pyparsing==3.0.9
# # python-dateutil==2.9.0.post0
# # python-json-logger==2.0.7
# # pytz==2022.6
# # PyYAML==6.0
# # pyzmq==26.2.0
# # referencing==0.35.1
# # regex==2024.9.11
# # requests==2.29.0
# # rfc3339-validator==0.1.4
# # rfc3986-validator==0.1.1
# # rpds-py==0.20.1
# # s3transfer==0.6.0
# # safetensors==0.4.5
# # scikit-image==0.24.0
# # scipy==1.9.3
# # SecretStorage==3.3.3
# # Send2Trash==1.8.3
# # sniffio==1.3.1
# # soupsieve==2.6
# # stack-data==0.6.3
# # sympy==1.13.1
# # terminado==0.18.1
# # tifffile==2024.8.30
# # tinycss2==1.4.0
# # tokenizers==0.20.2
# # tomli==2.0.2
# # torch==2.2.2
# # torchaudio==2.5.1
# # torchsummary==1.5.1
# # torchvision==0.17.2
# # tornado==6.4.1
# # tqdm==4.66.6
# # traitlets==5.14.3
# # transformers==4.46.1
# # triton==2.2.0
# # types-python-dateutil==2.9.0.20241003
# # uri-template==1.3.0
# # urllib3==1.26.16
# # wcwidth==0.2.13
# # webcolors==24.8.0
# # webencodings==0.5.1
# # websocket-client==1.8.0
# # widgetsnbextension==4.0.13
# # zipp==3.10.0
# # zstandard==0.23.0
# yt-dlp>=2024.11.18
# google-api-python-client
